Put tls test to dedicated target
